requirements.txt
djangorestframework
djangorestframework-jwt
django
django_postgrespool
psycopg2
clearbit
aiohttp
aiodns
cchardet
pyyaml
markdown
